requirements.txt
python-pptx>=0.6.21
pyinstaller>=5.0
pdf2image>=1.16.3
